Design and implement GolderCalculator for Hw2
`GoldenCalculator` with normal and random tests are designed and implement to be the software modeling of hardware `RobustCalculator`, which aim to be able to evaluate all kinds of legal expressions.

For testing, I design normal and random tests, the former is for basic and some edge cases, the latter use Python as golden in random generation and result verification.